VS2.25 - added storing starting_population counts
